+++ CHANGELOG.md
@@ -95,6 +95,8 @@
 - Enh #96: Add `HasTextDecoration` trait and corresponding tests for managing SVG `text-decoration` attribute (@terabytesoftw)
 - Enh #97: Add `HasWordSpacing` trait and corresponding tests for managing SVG `word-spacing` attribute (@terabytesoftw)
 - Bug #98: Update support descriptions for various SVG attributes to clarify assignment options (@terabytesoftw)
+- Enh #97: Add `HasTextLength` trait and corresponding tests for managing SVG `textLength` attribute (@terabytesoftw)
+- Enh #98: Add `HasWordSpacing` trait and corresponding tests for managing SVG `word-spacing` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024